requirements.txt
flask
pandas
pymongo 
python-dotenv
sentence-transformers 
requests
pinecone-client 
groq
